requirements.txt
customtkinter==5.2.2
pygame==2.5.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pillow==10.2.0
requests==2.31.0